移動 getPriceLimits() 的位置到 dbCompanies
將 getPriceLimits() 與輔助它的 isLowPriceCompany(), isHighPriceCompany() 從 helpers 移至 dbCompanies
