+++ db/dbCompanies.js
@@ -4,6 +4,7 @@ import { Mongo } from 'meteor/mongo';
 import SimpleSchema from 'simpl-schema';
 
 import { dbProducts } from './dbProducts';
+import { dbVariables } from './dbVariables';
 
 // 公司資料集
 export const dbCompanies = new Mongo.Collection('companies');
@@ -63,6 +64,41 @@ export function getUsedProductionFund(companyData) {
 // 取得公司剩餘可用的生產資金
 export function getAvailableProductionFund(companyData) {
   return getTotalProductionFund(companyData) - getUsedProductionFund(companyData);
+}
+
+// 判斷是否為低價公司
+export function isLowPriceCompany(companyData) {
+  const lowPriceThreshold = dbVariables.get('lowPriceThreshold');
+
+  return companyData.listPrice < lowPriceThreshold;
+}
+
+// 判斷是否為高價公司
+export function isHighPriceCompany(companyData) {
+  const highPriceThreshold = dbVariables.get('highPriceThreshold');
+
+  return companyData.listPrice >= highPriceThreshold;
+}
+
+// 取得買賣單的上下限
+export function getPriceLimits(companyData) {
+  const upper = getPriceUpperLimit(companyData);
+  const lower = getPriceLowerLimit(companyData);
+
+  return { upper, lower };
+}
+
+function getPriceUpperLimit(companyData) {
+  if (isLowPriceCompany(companyData)) {
+    return Math.ceil(companyData.listPrice * 1.30);
+  }
+  else {
+    return Math.ceil(companyData.listPrice * 1.15);
+  }
+}
+
+function getPriceLowerLimit(companyData) {
+  return Math.max(Math.floor(companyData.listPrice * 0.85), 1);
 }
 
 const schema = new SimpleSchema({
