+++ server/functions/company/helpers.js
@@ -1,41 +1,6 @@
 import { dbOrders } from '/db/dbOrders';
-import { dbVariables } from '/db/dbVariables';
 import { dbLog } from '/db/dbLog';
-
-// 判斷是否為低價公司
-export function isLowPriceCompany(companyData) {
-  const lowPriceThreshold = dbVariables.get('lowPriceThreshold');
-
-  return companyData.listPrice < lowPriceThreshold;
-}
-
-// 判斷是否為高價公司
-export function isHighPriceCompany(companyData) {
-  const highPriceThreshold = dbVariables.get('highPriceThreshold');
-
-  return companyData.listPrice >= highPriceThreshold;
-}
-
-// 取得買賣單的上下限
-export function getPriceLimits(companyData) {
-  const upper = getPriceUpperLimit(companyData);
-  const lower = getPriceLowerLimit(companyData);
-
-  return { upper, lower };
-}
-
-function getPriceUpperLimit(companyData) {
-  if (isLowPriceCompany(companyData)) {
-    return Math.ceil(companyData.listPrice * 1.30);
-  }
-  else {
-    return Math.ceil(companyData.listPrice * 1.15);
-  }
-}
-
-function getPriceLowerLimit(companyData) {
-  return Math.max(Math.floor(companyData.listPrice * 0.85), 1);
-}
+import { getPriceLimits } from '/db/dbCompanies';
 
 // 統計一段時間之內的交易數量
 export function calculateDealAmount(companyData, lookBackTime) {
